+++ project/src/Components.js
@@ -1,0 +1,6 @@
+// /home components
+export { default as NavBar } from "./components/NavBar";
+export { default as SelectList } from "./components/SelectList";
+export { default as SelectedList } from "./components/SelectedList";
+export { default as OptionList } from "./components/OptionList";
+export { default as MakeRequest } from "./components/MakeRequest";
